+++ src/main/java/com/shop/web/api/ProductController.java
@@ -11,6 +11,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.util.Collection;
+
 /**
  * Created by bymot on 21.12.2015.
  */
@@ -21,8 +23,8 @@ public class ProductController {
     private ProductService productService;
 
     @RequestMapping(value = "/api",
-    method = RequestMethod.GET,
-    produces = MediaType.APPLICATION_JSON_VALUE)
+            method = RequestMethod.GET,
+            produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<String> home() {
         return new ResponseEntity<String>("Hello world", HttpStatus.OK);
     }
@@ -30,10 +32,10 @@ public class ProductController {
     @RequestMapping(value = "/api/product/{id}",
             method = RequestMethod.GET,
             produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<Product> getProduct(@PathVariable("id") int id) {
+    public ResponseEntity<Product> getProduct(@PathVariable("id") Long id) {
         Product product = productService.findOne(id);
 
-        if(product == null) {
+        if (product == null) {
             return new ResponseEntity<Product>(HttpStatus.NOT_FOUND);
         }
 
@@ -41,4 +43,12 @@ public class ProductController {
     }
 
 
+//    @RequestMapping(value = "/api/products",
+//            method = RequestMethod.GET,
+//            produces = MediaType.APPLICATION_JSON_VALUE)
+//    public ResponseEntity<Collection<Product>> getAllProducts() {
+//        Collection<Product> products = productService.findAll();
+//
+//        return new ResponseEntity<Product>(products, HttpStatus.OK);
+//    }
 }
